pyproject.toml
[tool.poetry]
name = "pneumonia-classifier"
version = "0.1.0"
description = "End-to-end machine learning project for pneumonia classification from X-ray images"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pneumonia_classifier"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
torch = "^2.0.1"
torchvision = "^0.15.2"
fastapi = "^0.108.0"
uvicorn = "^0.22.0"
python-multipart = "^0.0.6"
pillow = "^9.5.0"
numpy = "^1.25.0"
pydantic = "^1.10.10"
scikit-learn = "^1.2.2"
mlflow = "2.10.2"
# Azure SDK dependencies
azure-identity = "^1.15.0"
azure-mgmt-resource = "^23.0.1"
azure-mgmt-containerinstance = "^10.1.0"
azure-mgmt-containerregistry = "^10.2.0"
httpx = "0.24.1"
setuptools = "^79.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.3.0"
flake8 = "^6.0.0"
pre-commit = "^3.3.2"
jupyter = "^1.0.0"
autoflake = "^2.3.1"
ruff = "^0.11.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"

[tool.poetry.scripts]
train = "scripts.train:main"
api = "scripts.api:main"
deploy = "scripts.deploy_to_azure:main"
build = "scripts.build_and_push:main" 